% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Add alpha (transparency) to a color}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{a color, specified as in \code{\link{colors}}, or in output format of \code{\link{rgb}}}

\item{alpha}{numeric vector w/ value(s) between 0 (transparent) and 1 (opaque)}
}
\description{
Friendly way to change the transparency of a color of vector of colors that are specified by name.
}
\details{
Pretty sure I got the code (or inspiration) \href{# http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html}{here}. This function is simple (the code is just 1 line), but it is quick and convenient.
}
\examples{
N <- 10
blues <- add.alpha("blue", seq(0.1,1, length.out=N))
plot(1:N, col=blues, pch=19)
b2r <- colorRampPalette

ba <- add.alpha("blue", alpha=0.25)
ra <- add.alpha("red", alpha=1)
cols <- colorRampPalette(c(ba, ra), alpha=TRUE)(N)
plot(1:N, col=cols, pch=19)
}
\seealso{
See \code{\link{rgb}}, \code{\link{colors}}, \code{\link{colorRampPalette}}, \code{\link{adjustcolor}} for base functions handling colors. See \code{\link{rbLib-package}} for package overview.
}

