% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pdf.R
\name{combine_pdf}
\alias{combine_pdf}
\title{Combine Multiple Files into a Single PDF}
\usage{
combine_pdf(inputNames, outputName)
}
\arguments{
\item{inputNames}{a character vector of files to combine; can include path}

\item{outputName}{the name of the merged pdf file}
}
\description{
Uses GhostScript to combine multiple files into a single pdf
}
\details{
Sets a few options, including the option to not rotate the images as they are combined. Only tested by me on OSX
}

