% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\title{Create mirror of a directory (copy)}
\usage{
mirror(from, to, exclude = "")
}
\arguments{
\item{from}{path of source directory}

\item{to}{path of destination directory}

\item{exclude}{character vector of files or folders to exclude; paths are relative}
}
\description{
\code{mirror} copies files from \code{from} to \code{to}, deleting any files in \code{to} that are not present in \code{from}.
}
\details{
Uses rsync. Essentially creates a copy of \code{from} at \code{to}.
}
\note{
When syncing the contents of a directory, remember that it is important to include a forward slash (\code{/}) at the end of both paths \code{a} and \code{b}; omitting in the second will create an extra copy in the first. If slash omitted from both, the top level name will exist 3 times and include two replicates. This is just basic \code{rsync} syntax, but it is easy to forget.
}
\examples{
\dontrun{
# I have a project in a folder called 'CyanoIso'
# I want two copies of this folder and its contents on my computer
from <- "Documents/School&Work/WiscResearch/CyanoIso/"
to <- "Dropbox/Cyanotopes/CyanoIso/" # omitting last slash creates extra copy
exclude <- c(".git", "Data")
(mirror(from, to, exclude))
}
}
\seealso{
See \code{\link{sync}} for bidirectional copying w/o deletions (syncing). See \code{\link{rbLib-package}} for package overview.
}

