% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renameNow.R
\name{renameNow}
\alias{renameNow}
\title{Append date-time to file string}
\usage{
renameNow(x)
}
\arguments{
\item{x}{one or more character strings ending with an extension (see Details for definition of extension)}
}
\description{
Add the date/time to a character string right before the file extension. Useful for adding the date to a default file name.
}
\details{
The file extension is defined by the regular expression pattern \code{"(\\.[^.]+$)"}. I.e., the extension is defined as the final period and the character(s) that follow it (not including end of string etc). Note that \code{"test.csv .,"} will be recognized as havign an extension (".,"), whereas \code{"test.csv ."} will return an error.
}
\examples{
renameNow(c("test1.csv","test2.txt"))
}
\seealso{
See \code{\link{rbLib-package}} for package overview.
}

