% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zCol.R
\name{zCol}
\alias{zCol}
\title{Z Color}
\usage{
zCol(nCols, Z)
}
\arguments{
\item{nCols}{number of colors}

\item{Z}{numeric vector to determine color}
}
\value{
a vector of colors

export
}
\description{
Get colors based on values
}

