% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prp.R
\name{prp}
\alias{prp}
\title{Push, Run, then Pull}
\usage{
prp(path, scriptName, remoteName, verbose = FALSE)
}
\arguments{
\item{path}{Character string for path to be pulled, or path containing file to be pulled}

\item{scriptName}{Character string of the script to be run}

\item{remoteName}{Character string for remote server in username@place.edu format. See \code{run.remote} in the package \code{ssh.utils}}

\item{verbose}{logical. print extra output.}
}
\description{
\code{prp} pushes local updates to remote, runs script remotely, then pulls remote updates to local
}
\examples{
\dontrun{
path <- "./Documents/School&Work/NCEAS_UnderIce/core/scripts/analysis/"
scriptName <- "reset.sim.R"
remoteName <- "ryanb@amphiprion.deenr.rutgers.edu"
prp(path, scriptName, remoteName, verbose=TRUE)
}
}
\seealso{
\code{\link{pull}} and \code{\link{push}} to sync files, and \code{\link{run}} to run script remotely.  See \code{\link{rbLib-package}} for overview.
}

