% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prp.R
\name{run}
\alias{run}
\title{Run a remote script remotely}
\usage{
run(scriptName, path, remoteName, verbose = FALSE)
}
\arguments{
\item{scriptName}{Character string indicating the name of the script to be run}

\item{path}{Character string for path to script. Must end with a "/"}

\item{remoteName}{Character string for remote server in username@place.edu format. See \code{\link{run.remote}} in the package \code{ssh.utils}}

\item{verbose}{logical. print extra output.}
}
\description{
\code{run} runs an R script remotely
}
\examples{
\dontrun{
	path <- "./Documents/School&Work/NCEAS_UnderIce/core/scripts/analysis/"
	scriptName <- "reset.sim.R"
	remoteName="ryanb@amphiprion.deenr.rutgers.edu"
	sync=TRUE
	(run(scriptName, path, remoteName))
}
}
\seealso{
\code{\link{pull}} for the pulling in remote files, \code{\link{push}} to push a local file to the remote server, and \code{\link{prp}} to push run pull. See \code{\link{rbLib-package}} for overview.
}

